\name{generateManifest}
\alias{generateManifest}
\docType{methods}
\title{
generateManifest
}
\description{
Generates a manifest file based on a list of entities objects.
To be removed in synapseclient v5.0.0. This is being replaced by `generate_manifest`.
}
\usage{
generateManifest(allFiles, filename, provenance_cache=NULL)
}
\arguments{
\item{allFiles}{ A list of File Entity objects on Synapse (can't be Synapse IDs)}
\item{filename}{ A file where manifest will be written}
\item{provenance_cache}{An optional dict of known provenance dicts keyed by entity ids}
}
\details{
% A detailed if possible precise description of the functionality provided, extending the basic information in the \description slot.
}
\value{

}
\references{
% A section with references to the literature. Use \url{} or \href{}{} for web pointers.
}
\note{
% Use this for a special note you want to have pointed out. 
}
\seealso{
% Pointers to related R objects, using \code{\link{...}} to refer to them.
}
% Examples of how to use the function. 
% Examples are not only useful for documentation purposes, but also provide test code used for diagnostic checking of R code. 
% By default, text will be displayed in the output of the help page and run by example() and by R CMD check. 
% You can use \dontrun{} for text that should only be shown, but not run, and \dontshow{} for extra commands 
% for testing that should not be shown to users, but will be run by example(). 
\examples{
}



